Added base for a new `Push` | `Pop` sequence variant to stack
Missing ArrayCodeDisplayBlock for user input
